Refactor Language Switcher widget. #206
