--- src/inc/feature.widget.php
@@ -1,3 +1,0 @@
-<?php # -*- coding: utf-8 -*-
-
-add_action( 'widgets_init', [ 'Mlp_Widget', 'widget_register' ] );
